chore: cleanup helpers to reduce code change
